ws/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
actuator-msgs==0.0.1
agent-interfaces==0.0.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
builtin-interfaces==1.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
composition-interfaces==1.2.1
conscience==0.0.0
control-msgs==4.6.0
controller-manager==2.42.0
controller-manager-msgs==2.42.0
cv-bridge==3.2.1
decorator==5.1.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
distro==1.9.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.2.2
executing==2.0.1
frozenlist==1.4.1
gazebo-msgs==3.7.0
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
greenlet==3.0.3
groq==0.10.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
image-geometry==3.2.1
interactive-markers==2.3.2
ipython==8.27.0
jedi==0.19.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-core==0.2.37
langchain-groq==0.1.9
langchain-text-splitters==0.2.2
langgraph==0.2.15
langgraph-checkpoint==1.0.8
langsmith==0.1.108
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
lxml==5.3.0
map-msgs==2.1.0
matplotlib-inline==0.1.7
message-filters==4.3.4
multidict==6.0.5
nav-msgs==4.2.4
numpy==1.26.4
orjson==3.10.7
osrf-pycommon==2.0.2
packaging==24.1
parso==0.8.4
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
python-qt-binding==1.1.2
PyYAML==6.0.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
requests==2.32.3
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.15
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2controlcli==2.42.0
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
scripts==3.7.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sros2==0.10.5
stack-data==0.6.3
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tenacity==8.5.0
tf2-geometry-msgs==0.25.7
tf2-kdl==0.25.7
tf2-msgs==0.25.7
tf2-py==0.25.7
tf2-ros-py==0.25.7
tf2-tools==0.25.7
topic-monitor==0.20.5
traitlets==5.14.3
trajectory-msgs==4.2.4
turtlesim==1.4.2
typing_extensions==4.12.2
unique-identifier-msgs==2.2.1
urllib3==2.2.2
vision-msgs==4.1.1
visualization-msgs==4.2.4
wcwidth==0.2.13
xacro==2.0.8
yarl==1.9.6
